pyproject.toml
[project]
name = "auto-comply"
version = "0.1.0"
description = "AI agent for completing your workplace training courses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mss>=9.0.2",
    "numpy>=2.1.3",
    "openai>=1.54.3",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "pyobjc-framework-quartz>=10.3.1",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.5.1",
]
